pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
molecule-plugins = {extras = ["docker"], version = "^23.7.0"}
ansible-core = "^2.18.4"
molecule = "^25.4.0"
requests = "^2.32.3"
dnspython = "^2.7.0"
